+++ CHANGELOG.md
@@ -6,10 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - PASSBOLT-2978: Open a paginated grid on a page containing a target item
 - PASSBOLT-3285: The url should be updated when the user is selecting a password
-- PASSBOLT-2995 As LU I should be able to copy the permalink of a password
+- PASSBOLT-2995: As LU I should be able to copy the permalink of a password
 - PASSBOLT-3312: As GM adding a user to a group I should see a relevant feedback in case of network/proxy errors
 - PASSBOLT-3318: As LU I should retrieve a secret when I'm copying it
 - PASSBOLT-3319: As LU I should retrieve a secret when I'm editing it
+- PASSBOLT-3403: As LU I should retrieve secrets when I'm exporting the associated passwords
+- PASSBOLT-3397: Remove the list of secrets from the API request while loading the list of passwords
 
 ### Fixed
 - PASSBOLT-3268: BaseDN should not be mandatory